feat(client): extract completion context detection into kaish_client::completion
The kaish-extras browser playground is growing tab completion and needs
the same first answer the rustyline REPL computes on Tab: what kind of
thing is under the cursor. That's pure text analysis, so it moves to
kaish-client — one detector, every frontend — with the REPL now consuming
it. Decisions from the kaish-extras session with Amy:

- Detection only moves; candidate sourcing stays frontend-local. The REPL
  keeps rustyline's FilenameCompleter for real-fs paths; the web frontend
  completes against the kernel VFS. Forcing one candidate pipeline would
  have dragged rustyline types into the client crate.
- CompletionContext gains Debug/PartialEq derives (tests assert directly),
  plus a word_start() helper for the start-of-word scan both frontends
  repeat.
- EmbeddedClient's blob-id timestamp now goes through kaish_types::clock,
  finishing the browser-safety audit item from #205 — kaish-client is now
  compilable-and-callable on wasm32-unknown-unknown.

Detection tests move to the new module (matches! upgraded to assert_eq);
client 19 + repl 39+4 green, clippy clean.
