+++ crates/kaish-client/Cargo.toml
@@ -14,6 +14,7 @@ readme = "../../README.md"
 
 [dependencies]
 kaish-kernel = { path = "../kaish-kernel", version = "0.12.0" }
+kaish-types = { path = "../kaish-types", version = "0.12.0" }
 
 # Async runtime
 tokio = { workspace = true }
